Update backup-and-restore.md
Based on the latest case reviews, customers asked about long term backup (LTB) concept. in Azure Synapse SQL Pool (LTB) is not supported so we are advising the customers to restore a copy from the Datawarehouse and pause it for long term.

The above information's available in this document, but since the note contains two parts, the first one related to vote for the feature and the second one about the LTR concepts it seems our customers missing that part.

The proposed change is to split the note into two parts and identify the steps which the customer should perform. Thanks